requirements.txt
django-taggit-autosuggest
install django-mathjax
